Struct cluu_kernel_rust::arch::x86_64::peripheral::uart_16550::SerialPort

pub struct SerialPort<T: Io> {
    data: T,
    int_en: T,
    fifo_ctrl: T,
    line_ctrl: T,
    modem_ctrl: T,
    line_sts: ReadOnly<T>,
    modem_sts: ReadOnly<T>,
}

Serial port representation.

Fields
data: Tint_en: Tfifo_ctrl: Tline_ctrl: Tmodem_ctrl: Tline_sts: ReadOnly<T>modem_sts: ReadOnly<T>
Implementations

impl SerialPort<Pio<u8>>

pub const fn new(base: u16) -> SerialPort<Pio<u8>>

Creates a new serial port instance.

Arguments

base - The base port address of the serial port.

Returns

Returns a new SerialPort instance.

impl<T: Io> SerialPort<T>where T::Value: From<u8> + TryInto<u8>,

pub fn init(&mut self)

Initializes the serial port.

fn line_sts(&self) -> LineStsFlags

Retrieves the line status flags.

Returns

Returns the line status flags indicating the current status of the serial port.

pub fn send(&mut self, data: u8)

Sends a byte of data through the serial port.

Arguments

data - The data byte to send.

pub fn write(&mut self, b: u8)

Writes a byte of data to the serial port.

Arguments

b - The byte of data to write.