Trait cluu_kernel_rust::syscall::io::Io

pub trait Io {
    type Value: Copy + PartialEq + BitAnd<Output = Self::Value> + BitOr<Output = Self::Value> + Not<Output = Self::Value>;

    // Required methods
    fn read(&self) -> Self::Value;
    fn write(&mut self, value: Self::Value);

    // Provided methods
    fn readf(&self, flags: Self::Value) -> bool { ... }
    fn writef(&mut self, flags: Self::Value, value: bool) { ... }
}

Represents an I/O interface.

Required Associated Types

type Value: Copy + PartialEq + BitAnd<Output = Self::Value> + BitOr<Output = Self::Value> + Not<Output = Self::Value>

The value type used for I/O operations.

Required Methods

fn read(&self) -> Self::Value

Reads the value from the I/O interface.

fn write(&mut self, value: Self::Value)

Writes the value to the I/O interface.

Provided Methods

fn readf(&self, flags: Self::Value) -> bool

Reads the value from the I/O interface and checks if the specified flags are set.

Arguments

flags - The flags to check.

Returns

Returns true if all the specified flags are set, false otherwise.

fn writef(&mut self, flags: Self::Value, value: bool)

Writes the value to the I/O interface based on the specified flags and value.

Arguments

flags - The flags to modify.

value - The value indicating whether to set (true) or clear (false) the flags.